+++ src/scvi/autotune/_tune.py
@@ -31,6 +31,9 @@ def run_autotune(
     logging_dir: str | None = None,
     scheduler_kwargs: dict | None = None,
     searcher_kwargs: dict | None = None,
+    scib_stage: str | None = "train_end",
+    scib_subsample_rows: int | None = 5000,
+    scib_indices_list: list | None = None,
 ) -> AutotuneExperiment:
     """``BETA`` Run a hyperparameter sweep.
 
@@ -95,6 +98,15 @@ def run_autotune(
         Additional keyword arguments to pass to the scheduler.
     searcher_kwargs
         Additional keyword arguments to pass to the search algorithm.
+    scib_stage
+        Used when performing scib-metrics tune, select whether to perform on "validation_end"
+        or "train_end" (default).
+    scib_subsample_rows
+        Used when performing scib-metrics tune, select number of rows to subsample (5000 default).
+        This is important to save computation time
+    scib_indices_list
+        If not empty will be used to select the indices to calc the scib metric on, otherwise will
+        use the random indices selection in size of scib_subsample_rows
 
     Returns
     -------
@@ -112,12 +124,12 @@ def run_autotune(
     from ray import init
 
     experiment = AutotuneExperiment(
-        model_cls,
-        data,
-        metrics,
-        mode,
-        search_space,
-        num_samples,
+        model_cls=model_cls,
+        data=data,
+        metrics=metrics,
+        mode=mode,
+        search_space=search_space,
+        num_samples=num_samples,
         scheduler=scheduler,
         searcher=searcher,
         seed=seed,
@@ -126,8 +138,11 @@ def run_autotune(
         logging_dir=logging_dir,
         scheduler_kwargs=scheduler_kwargs,
         searcher_kwargs=searcher_kwargs,
+        scib_stage=scib_stage,
+        scib_subsample_rows=scib_subsample_rows,
+        scib_indices_list=scib_indices_list,
     )
     logger.info(f"Running autotune experiment {experiment.name}.")
-    init(log_to_driver=False, ignore_reinit_error=True)
+    init(log_to_driver=False, ignore_reinit_error=True, local_mode=True)
     experiment.result_grid = experiment.get_tuner().fit()
     return experiment