+++ tests/autotune/test_tune.py
@@ -1,10 +1,12 @@
+import pytest
 from ray import tune
 from ray.tune import ResultGrid
 
 from scvi import settings
 from scvi.autotune import AutotuneExperiment, run_autotune
 from scvi.data import synthetic_iid
-from scvi.model import SCVI
+from scvi.dataloaders import DataSplitter
+from scvi.model import SCANVI, SCVI
 
 
 def test_run_autotune_scvi_basic(save_path: str):
@@ -25,7 +27,7 @@ def test_run_autotune_scvi_basic(save_path: str):
                 "max_epochs": 1,
             },
         },
-        num_samples=1,
+        num_samples=2,
         seed=0,
         scheduler="asha",
         searcher="hyperopt",
@@ -36,8 +38,6 @@ def test_run_autotune_scvi_basic(save_path: str):
 
 
 def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int = 3):
-    from scvi.dataloaders import DataSplitter
-
     settings.logging_dir = save_path
     adata = synthetic_iid(n_batches=n_batches)
     SCVI.setup_anndata(adata, batch_key="batch")
@@ -49,7 +49,7 @@ def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int = 3):
 
     experiment = run_autotune(
         SCVI,
-        datamodule,
+        data=datamodule,
         metrics=["elbo_validation"],
         mode="min",
         search_space={
@@ -60,7 +60,50 @@ def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int = 3):
                 "max_epochs": 1,
             },
         },
-        num_samples=1,
+        num_samples=2,
+        seed=0,
+        scheduler="asha",
+        searcher="hyperopt",
+    )
+    assert isinstance(experiment, AutotuneExperiment)
+    assert hasattr(experiment, "result_grid")
+    assert isinstance(experiment.result_grid, ResultGrid)
+
+
+@pytest.mark.parametrize("metric", ["Total", "Bio conservation", "iLISI"])
+@pytest.mark.parametrize("model_cls", [SCVI, SCANVI])
+def test_run_autotune_scvi_with_scib(model_cls, metric: str, save_path: str):
+    settings.logging_dir = save_path
+    adata = synthetic_iid()
+    if model_cls == SCANVI:
+        model_cls.setup_anndata(
+            adata,
+            labels_key="labels",
+            unlabeled_category="unknown",
+            batch_key="batch",
+        )
+    else:
+        model_cls.setup_anndata(
+            adata,
+            labels_key="labels",
+            batch_key="batch",
+        )
+
+    experiment = run_autotune(
+        model_cls,
+        adata,
+        metrics=[metric],
+        mode="max",
+        search_space={
+            "model_params": {
+                "n_hidden": tune.choice([1, 2]),
+            },
+            "train_params": {
+                "max_epochs": 1,
+            },
+        },
+        num_samples=2,
+        scib_subsample_rows=100,
         seed=0,
         scheduler="asha",
         searcher="hyperopt",