+++ CHANGELOG.md
@@ -21,6 +21,8 @@ to [Semantic Versioning]. Full commit history is available in the
     AutoZI, CellAssign and GimVI. {pr}`3121`.
 - Add {class}`scvi.external.RESOLVI` for bias correction in single-cell resolved spatial
     transcriptomics {pr}`3144`.
+- Add scib-metrics support for {class}`scvi.autotune.AutotuneExperiment` and
+    {class}`scvi.train._callbacks.ScibCallback` for autotune for scib metrics {pr}`3168`.
 
 #### Fixed